Task 1C First Commit
Firstly imported the Task 1C unity project files, this included the enemies, bullets etc... I then added in the spawn over time script on top of the spawner 'enemy' spawner. This one spawns the pickups for the player. Then changes made within the prefabs now upon collision with the player and pickup, the pickup is destroyed and not yet but will soon change the players weapon. The enemyboss now spawns however problems around the movement and bullet spawns plus speed of the enemyboss are still there.