Fixing the change in weapon upon pickup
So during this time i want to say I learnt more about what was happening with the code between different scripts. I learnt what was needed inside the pickup and the player. At the end of this the player starts with the 'machine gun' and changes to the triple shot when colliding with the pickup