Machine Gun & Triple Shot Work
Currently the machine gun works upon game start, the weapon fires at the specified speed and can be changed based on the value in the delay slot inside of the spaceship inspector. The weapon changes to the triple shot upon collision with the pickup and the triple shot now fires three bullets with a delay before being able to fire again. However the bullet spread is large and not straight up, plus each weapon when fired moves the spaceship slightly to the right