EnemyShip, Boss and weapon styles all working
I managed to get the enemies spawning and firing working fine. Same with the boss which spawns and fires the triple shot which works well. To fix the bullets going upwards instead of down I used a serialize field for a float value to determine the direciton of the bullets. Preset at 0.5f which what it was before however for the boss I manually changed it to -0.5f. All collision works as intended now too.