Final Commit
Within this change I managed to combine the Playermovement script and the enemymovement script into one, i.e the EngineBase script. Just by replacing some names for example, the playerinput called to the playermovement but now calls the EngineBase. The player also now doesn't displace upon firing which was an issue before.

The game is now playable where the player is controllable, the weapon changes upon collision with the pickup from the machine gun to the triple shot. The enemy ships spawn and fire, movement works and same goes for the boss.